Revise points functionality.

+++ app/src/main/java/com/example/vjoshi/wattsapp/addDeviceClasses/DeviceConstants.java
@@ -1,8 +1,8 @@
 package com.example.vjoshi.wattsapp.addDeviceClasses;
 
 public interface DeviceConstants {
-    long USAGE_THRESHOLD = 50000;
-    long USAGE_STEP = 1000;
+    double USAGE_THRESHOLD = 50000;
+    double USAGE_STEP = 1000;
 
     String TAG = "TAG";
     String DEVICENAME = "device";
